Rename exporter.go
